requirements.txt
django==4.0.6
Pillow==9.2.0
django-phonenumber-field==6.3.0
environs==9.5.0
phonenumbers==8.12.52
flake8==5.0.3
python-telegram-bot==13.13